Made some refactoring
