--- BasicChatContents/BasicChatContents/Message.cs
@@ -1,16 +1,0 @@
-﻿using System;
-
-namespace BasicChatContents
-{
-    [Serializable]
-    public class Message<T>
-    {
-        public Guid ChatId { get; set; }
-        public T MessageToSend { get; set; }
-
-        public override string ToString()
-        {
-            return MessageToSend.ToString();
-        }
-    }
-}
