+++ ChatClient/ChatClient/Chats/BasicChat.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using BasicChatContents;
+using System;
 using System.Collections.Generic;
 using System.Net.Sockets;
 using System.Text;
