+++ ChatServer/ChatServer/ChatManagers/IChatManager.cs
@@ -8,7 +8,7 @@ namespace ChatServer.ChatManagers
         public List<User> OtherUsersInChat { get; set; }
         void RemoveClientFromReceivingMessages(User user);
         void EnterUserToChat(User user);
-
-
+        void ChatWithClient(User user);
+        //void SendMessageToClients(string dataToSend);
     }
 }
