--- ChatServer/ChatServer/ConsoleDisplayer.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace ChatServer
-{
-    public class ConsoleDisplayer
-    {
-        public void PrintValueToConsole(object obj)
-        {
-            Console.WriteLine(obj);
-        }
-    }
-}
